pyproject.toml
[tool.poetry]
name = "livevil_blog"
version = "0.1.0"
description = "blog powered by django"
authors = ["xlivevil <xlivevil@aliyun.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
django-contrib-comments = "^2.1.0"
django-haystack = "^3.0"
django-allauth = "^0.44.0"
requests = "^2.25.1"
Pillow = "^8.2.0"
djangorestframework = "^3.12.4"
django-crispy-forms = "^1.11.2"
django-reversion = "^3.0.9"
mysqlclient = "^2.0.3"
django-redis = "^4.12.1"
django-filter = "^2.4.0"
django-mptt = "^0.12.0"
subdomains = "^3.0.1"
drf-haystack = "^1.8.10"
django-grappelli = "^2.15.1"
drf-yasg = "^1.20.0"
django-compressor = "^2.4.1"
daphne = "^3.0.2"
gunicorn = "^20.1.0"
elasticsearch = ">=5,<6"
sqlparse = "0.2.4"
graphene-django = "^2.15.0"
django-contact-form = {extras = ["akismet"], version = "^1.8.2"}
django-debug-toolbar = "^3.2.1"
djongo = "^1.3.6"
django-pure-pagination = "^0.3.0"
Markdown = "^3.3.4"
docutils = "^0.17.1"
drf-extensions = "^0.7.0"
sentry-sdk = "^1.1.0"
ibm-cos-sdk = "^2.10.0"
djangorestframework-simplejwt = "^4.8.0"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]
Faker = "^8.1.4"
fabric = "^2.6.0"
invoke = "^1.5.0"
mypy = "^0.812"
flake8 = "^3.9.2"
yapf = "^0.31.0"
coverage = {extras = ["toml"], version = "^5.5"}
isort = "^5.8.0"
django-stubs = "^1.8.0"
factory-boy = "^3.2.0"
djangorestframework-stubs = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
skip=[".tox", "migrations"]
multi_line_output="3"
known_third_party="django"
include_trailing_comma = "true"

[tool.flake8]
exclude = "*migrations*"
max-line-length = "119"